move init audio state
